UI: Fix service-worker.js not being updated
The commit 0e0df6fd changed the front-end build process to keep the
assets of the previous builds in the ui/dist directory (instead of
deleting all files in the dist folder, which is the default behavior of
Vite). Since the build of the service-worker didn't compress its output
file, the previously built service-worker.js file was being compressed
by subsequent runs of `npm run build`; and so serving the clients who
support HTTP compression an outdated version of the 'service-worker.js'
file. This commit fixes that.